Bugfix
Some bugfixes with complete new system for spawning and deleting the crystals. The crystals will now be removed if the plugin disables and spawning complete new if its loading. Also added the feature to save the locations of the crystals and not only the UUID and the GSON library is now not included to the plugin anymore. GSON lib will be exportet to s.Tools. s.Tools will be auto installed if you start the plugin without s.Tools.